Move data.ts back to webapp
