+++ webapp/src/auth.ts
@@ -2,15 +2,34 @@ import * as core from "./core";
 import * as data from "./data";
 import * as cloud from "./cloud";
 
+/**
+ * Virtual API keys
+ */
+const MODULE = "auth";
+const FIELD_USER_PROFILE = "profile";
+const FIELD_LOGGED_IN = "logged-in";
+export const USER_PROFILE = `${MODULE}:${FIELD_USER_PROFILE}`;
+export const LOGGED_IN = `${MODULE}:${FIELD_LOGGED_IN}`;
+
+const USER_PREF_MODULE = "user-pref";
+const FIELD_USER_PREFERENCES = "preferences";
+const FIELD_HIGHCONTRAST = "high-contrast";
+const FIELD_LANGUAGE = "language";
+const FIELD_READER = "reader";
+export const USER_PREFERENCES = `${USER_PREF_MODULE}:${FIELD_USER_PREFERENCES}`
+export const HIGHCONTRAST = `${USER_PREF_MODULE}:${FIELD_HIGHCONTRAST}`
+export const LANGUAGE = `${USER_PREF_MODULE}:${FIELD_LANGUAGE}`
+export const READER = `${USER_PREF_MODULE}:${FIELD_READER}`
+
 export class Component<TProps, TState> extends data.Component<TProps, TState> {
     public getUserProfile(): pxt.auth.UserProfile {
-        return this.getData<pxt.auth.UserProfile>(pxt.auth.USER_PROFILE);
+        return this.getData<pxt.auth.UserProfile>(USER_PROFILE);
     }
     public getUserPreferences(): pxt.auth.UserPreferences {
-        return this.getData<pxt.auth.UserPreferences>(pxt.auth.USER_PREFERENCES);
+        return this.getData<pxt.auth.UserPreferences>(USER_PREFERENCES);
     }
     public isLoggedIn(): boolean {
-        return this.getData<boolean>(pxt.auth.LOGGED_IN);
+        return this.getData<boolean>(LOGGED_IN);
     }
 }
 
@@ -28,13 +47,101 @@ class AuthClient extends pxt.auth.AuthClient {
         core.errorNotification(lf("Sign in failed. Something went wrong."));
         return Promise.resolve();
     }
+    protected onUserProfileChanged(): Promise<void> {
+        const state = this.getState();
+        generateUserProfilePicDataUrl(state.profile);
+        data.invalidate("auth:*");
+        return Promise.resolve();
+    }
+    protected onUserPreferencesChanged(part: pxt.auth.UserPreferencePart): Promise<void> {
+        switch (part) {
+            case "language": data.invalidate(LANGUAGE); break;
+            case "high-contrast": data.invalidate(HIGHCONTRAST); break;
+            case "immersive-reader": data.invalidate(READER); break;
+        }
+        return Promise.resolve();
+    }
     protected async onProfileDeleted(userId: string): Promise<void> {
         // Convert cloud-saved projects to local projects.
         await cloud.convertCloudToLocal(userId);
     }
+    protected onStateLoaded(): Promise<void> {
+        const state = this.getState();
+        generateUserProfilePicDataUrl(state.profile);
+        data.invalidate("auth:*");
+        data.invalidate("user-pref:*");
+        return Promise.resolve();
+    }
     protected onApiError(err: any): Promise<void> {
         core.handleNetworkError(err);
         return Promise.resolve();
+    }
+    protected onStateCleared(): Promise<void> {
+        data.invalidate("auth:*");
+        //data.invalidate("user-prefs:*"); // Should we invalidate this? Or would it be jarring visually?
+        return Promise.resolve();
+    }
+
+    public static authApiHandler(p: string): pxt.auth.UserProfile | boolean | string {
+        const cli = client();
+        if (cli) {
+            const field = data.stripProtocol(p);
+            const state = cli.getState();
+            switch (field) {
+                case FIELD_USER_PROFILE: return { ...state.profile };
+                case FIELD_LOGGED_IN: return cli.hasUserId();
+            }
+        }
+        return null;
+    }
+
+    public static userPreferencesHandler(path: string): pxt.auth.UserPreferences | boolean | string {
+        const cli = client();
+        if (cli) {
+            const state = cli.getState();
+            if (!state.preferences) {
+                cli.initialUserPreferencesAsync().then(() => { });
+            }
+            return AuthClient.internalUserPreferencesHandler(path);
+        }
+        return null;
+    }
+
+    private static internalUserPreferencesHandler(path: string): pxt.auth.UserPreferences | boolean | string {
+        const cli = client();
+        if (cli) {
+            const state = cli.getState();
+            const field = data.stripProtocol(path);
+            switch (field) {
+                case FIELD_USER_PREFERENCES: return { ...state.preferences };
+                case FIELD_HIGHCONTRAST: return state.preferences?.highContrast ?? pxt.auth.DEFAULT_USER_PREFERENCES().highContrast;
+                case FIELD_LANGUAGE: return state.preferences?.language ?? pxt.auth.DEFAULT_USER_PREFERENCES().language;
+                case FIELD_READER: return state.preferences?.reader ?? pxt.auth.DEFAULT_USER_PREFERENCES().reader;
+            }
+            return state.preferences
+        }
+        return null;
+    }
+}
+
+export function initVirtualApi() {
+    data.mountVirtualApi(USER_PREF_MODULE, {
+        getSync: AuthClient.userPreferencesHandler,
+    });
+    data.mountVirtualApi(MODULE, {
+        getSync: AuthClient.authApiHandler
+    });
+}
+
+function generateUserProfilePicDataUrl(profile: pxt.auth.UserProfile) {
+    if (profile?.idp?.picture?.encoded) {
+        const url = window.URL || window.webkitURL;
+        try {
+            // Decode the base64 image to a data URL.
+            const decoded = pxt.Util.stringToUint8Array(atob(profile.idp.picture.encoded));
+            const blob = new Blob([decoded], { type: profile.idp.picture.mimeType });
+            profile.idp.picture.dataUrl = url.createObjectURL(blob);
+        } catch { }
     }
 }
 
@@ -48,15 +155,15 @@ export function hasIdentity(): boolean {
 }
 
 export function loggedIn(): boolean {
-    return pxt.data.getData<boolean>(pxt.auth.LOGGED_IN);
+    return data.getData<boolean>(LOGGED_IN);
 }
 
 export function userProfile(): pxt.auth.UserProfile {
-    return pxt.data.getData<pxt.auth.UserProfile>(pxt.auth.USER_PROFILE);
+    return data.getData<pxt.auth.UserProfile>(USER_PROFILE);
 }
 
 export function userPreferences(): pxt.auth.UserPreferences {
-    return pxt.data.getData<pxt.auth.UserPreferences>(pxt.auth.USER_PREFERENCES);
+    return data.getData<pxt.auth.UserPreferences>(USER_PREFERENCES);
 }
 
 export function authCheckAsync(): Promise<pxt.auth.UserProfile | undefined> {
