+++ webapp/src/data.ts
@@ -4,14 +4,255 @@ import * as core from "./core";
 import Cloud = pxt.Cloud;
 import Util = pxt.Util;
 
-pxt.data.mountVirtualApi("cloud", {
-    getAsync: p => Cloud.privateGetAsync(pxt.data.stripProtocol(p)).catch(core.handleNetworkError),
+
+export type Action = () => void;
+export type DataSubscriber = {
+    subscriptions: CacheEntry[];
+    onDataChanged: (path: string) => void;
+};
+
+export interface CacheEntry {
+    path: string;
+    data: any;
+    lastRefresh: number;
+    queued: boolean;
+    callbackOnce: Action[];
+    components: DataSubscriber[];
+    api: VirtualApi;
+}
+
+export enum FetchStatus {
+    Pending,
+    Complete,
+    Error,
+    Offline
+};
+
+export interface DataFetchResult<T> {
+    data?: T;
+    status: FetchStatus;
+}
+
+const virtualApis: pxt.Map<VirtualApi> = {};
+let cachedData: pxt.Map<CacheEntry> = {};
+
+export function subscribe(component: DataSubscriber, path: string) {
+    let e = lookup(path)
+    let lst = e.components
+    if (lst.indexOf(component) < 0) {
+        lst.push(component)
+        component.subscriptions.push(e)
+    }
+}
+
+export function unsubscribe(component: DataSubscriber) {
+    let lst = component.subscriptions
+    if (lst.length == 0) return
+    component.subscriptions = []
+    lst.forEach(ce => {
+        let idx = ce.components.indexOf(component)
+        if (idx >= 0) ce.components.splice(idx, 1)
+    })
+}
+
+function expired(ce: CacheEntry) {
+    if (!ce.api.expirationTime)
+        return !ce.lastRefresh; // needs to be refreshed at least once
+    return ce.data == null || (Date.now() - ce.lastRefresh) > ce.api.expirationTime(ce.path)
+}
+
+function shouldCache(ce: CacheEntry) {
+    if (!ce.data || ce.data instanceof Error) return false;
+    return /^cloud:(me\/settings|ptr-pkg-)/.test(ce.path);
+}
+
+export function clearCache() {
+    cachedData = {};
+    saveCache();
+}
+
+export function loadCache() {
+    JSON.parse(pxt.storage.getLocal("apiCache2") || "[]").forEach((e: any) => {
+        let ce = lookup(e.path)
+        ce.data = e.data
+    });
+}
+
+function saveCache() {
+    let obj = Util.values(cachedData).filter(e => shouldCache(e)).map(e => {
+        return {
+            path: e.path,
+            data: e.data
+        }
+    });
+    pxt.storage.setLocal("apiCache2", JSON.stringify(obj));
+}
+
+function matches(ce: CacheEntry, prefix: string) {
+    return ce.path.slice(0, prefix.length) == prefix;
+}
+
+function notify(ce: CacheEntry, path: string) {
+    if (shouldCache(ce)) saveCache();
+
+    let lst = ce.callbackOnce
+    if (lst.length > 0) {
+        ce.callbackOnce = []
+        Util.nextTick(() => lst.forEach(f => f()))
+    }
+
+    if (ce.components.length > 0)
+        ce.components.forEach(c => Util.nextTick(() => c.onDataChanged(path)))
+}
+
+function getVirtualApi(path: string) {
+    let m = /^([\w\-]+):/.exec(path)
+    if (!m || !virtualApis[m[1]])
+        Util.oops("bad data protocol: " + path)
+    return virtualApis[m[1]]
+}
+
+function queueNotify(ce: CacheEntry, path: string) {
+    if (ce.queued) return
+
+    if (ce.api.isOffline && ce.api.isOffline())
+        return
+
+    ce.queued = true
+
+    let final = (res: any) => {
+        ce.data = res
+        ce.lastRefresh = pxt.Util.now()
+        ce.queued = false
+        notify(ce, path)
+    }
+
+    if (ce.api.isSync)
+        final(ce.api.getSync(ce.path))
+    else {
+        const p = ce.api.getAsync(ce.path);
+        p.then(final)
+    }
+}
+
+function lookup(path: string) {
+    if (!cachedData.hasOwnProperty(path))
+        cachedData[path] = {
+            path: path,
+            data: null,
+            lastRefresh: 0,
+            queued: false,
+            callbackOnce: [],
+            components: [],
+            api: getVirtualApi(path)
+        }
+    return cachedData[path]
+}
+
+export function getCached(component: DataSubscriber, path: string): DataFetchResult<any> {
+    subscribe(component, path)
+    return getDataWithStatus(path);
+}
+
+//
+// Public interface
+//
+
+export interface VirtualApi {
+    getAsync?(path: string): Promise<any>;
+    getSync?(path: string): any;
+    isSync?: boolean;
+    expirationTime?(path: string): number; // in milliseconds
+    isOffline?: () => boolean;
+    onInvalidated?: (path: string) => void;
+}
+
+export function mountVirtualApi(protocol: string, handler: VirtualApi) {
+    Util.assert(!virtualApis[protocol])
+    Util.assert(!!handler.getSync || !!handler.getAsync)
+    Util.assert(!!handler.getSync != !!handler.getAsync)
+    handler.isSync = !!handler.getSync
+    virtualApis[protocol] = handler
+}
+
+export function stripProtocol(path: string) {
+    let m = /^([\w\-]+):(.*)/.exec(path)
+    if (m) return m[2]
+    else Util.oops("protocol missing in: " + path)
+    return path
+}
+
+export function invalidate(path: string) {
+    const prefix = path.replace(/:\*$/, ':'); // remove trailing "*";
+    Util.values(cachedData).forEach(ce => {
+        if (matches(ce, prefix)) {
+            ce.lastRefresh = 0;
+            if (ce.components.length > 0)
+                queueNotify(lookup(ce.path), path)
+            if (ce.api.onInvalidated) {
+                ce.api.onInvalidated(path);
+            }
+        }
+    })
+}
+
+export function getAsync<T = any>(path: string) {
+    let ce = lookup(path)
+
+    if (ce.api.isSync)
+        return Promise.resolve(ce.api.getSync(ce.path))
+
+    if (!Cloud.isOnline() || !expired(ce))
+        return Promise.resolve(ce.data)
+
+    return new Promise<T>((resolve, reject) => {
+        ce.callbackOnce.push(() => {
+            resolve(ce.data)
+        })
+        queueNotify(ce, path)
+    })
+}
+
+export function getData<T>(path: string): T {
+    return getDataWithStatus<T>(path).data;
+}
+
+export function getDataWithStatus<T>(path: string): DataFetchResult<T> {
+    let r = lookup(path)
+    if (r.api.isSync)
+        return {
+            data: r.api.getSync(r.path),
+            status: FetchStatus.Complete
+        }
+
+    // cache async values
+    let fetchRes: DataFetchResult<T> = {
+        data: r.data,
+        status: FetchStatus.Complete
+    };
+
+    if (expired(r) || r.data instanceof Error) {
+        fetchRes.status = r.data instanceof Error ? FetchStatus.Error : FetchStatus.Pending;
+        if (r.api.isOffline && r.api.isOffline()) {
+            // The request will not be requeued so we don't want to show it as pending
+            fetchRes.status = FetchStatus.Offline;
+        } else {
+            queueNotify(r, path)
+        }
+    }
+
+    return fetchRes;
+}
+
+
+mountVirtualApi("cloud", {
+    getAsync: p => Cloud.privateGetAsync(stripProtocol(p)).catch(core.handleNetworkError),
     expirationTime: p => 60 * 1000,
     isOffline: () => !Cloud.isOnline(),
 })
 
-pxt.data.mountVirtualApi("cloud-search", {
-    getAsync: p => Cloud.privateGetAsync(pxt.data.stripProtocol(p)).catch(e => {
+mountVirtualApi("cloud-search", {
+    getAsync: p => Cloud.privateGetAsync(stripProtocol(p)).catch(e => {
         core.handleNetworkError(e, [404])
         return { statusCode: 404, headers: {}, json: {} }
     }),
@@ -19,33 +260,33 @@ pxt.data.mountVirtualApi("cloud-search", {
     isOffline: () => !Cloud.isOnline(),
 })
 
-pxt.data.mountVirtualApi("gallery", {
-    getAsync: p => pxt.gallery.loadGalleryAsync(pxt.data.stripProtocol(decodeURIComponent(p))).catch((e) => {
+mountVirtualApi("gallery", {
+    getAsync: p => pxt.gallery.loadGalleryAsync(stripProtocol(decodeURIComponent(p))).catch((e) => {
         return Promise.resolve(e);
     }),
     expirationTime: p => 3600 * 1000
 })
 
-pxt.data.mountVirtualApi("gh-search", {
+mountVirtualApi("gh-search", {
     getAsync: query => pxt.targetConfigAsync()
-        .then(config => pxt.github.searchAsync(pxt.data.stripProtocol(query), config ? config.packages : undefined))
+        .then(config => pxt.github.searchAsync(stripProtocol(query), config ? config.packages : undefined))
         .catch(core.handleNetworkError),
     expirationTime: p => 60 * 1000,
     isOffline: () => !Cloud.isOnline(),
 })
 
-pxt.data.mountVirtualApi("gh-pkgcfg", {
+mountVirtualApi("gh-pkgcfg", {
     getAsync: query =>
-        pxt.github.pkgConfigAsync(pxt.data.stripProtocol(query)).catch(core.handleNetworkError),
+        pxt.github.pkgConfigAsync(stripProtocol(query)).catch(core.handleNetworkError),
     expirationTime: p => 60 * 1000,
     isOffline: () => !Cloud.isOnline(),
 })
 
 // gh-commits:repo#sha
-pxt.data.mountVirtualApi("gh-commits", {
+mountVirtualApi("gh-commits", {
     getAsync: query => {
-        const p = pxt.data.stripProtocol(query);
-        const [ repo, sha ] = p.split('#', 2)
+        const p = stripProtocol(query);
+        const [repo, sha] = p.split('#', 2)
         return pxt.github.getCommitsAsync(repo, sha).catch(e => {
             core.handleNetworkError(e);
             return [];
@@ -56,16 +297,16 @@ pxt.data.mountVirtualApi("gh-commits", {
 })
 
 let targetConfigPromise: Promise<pxt.TargetConfig> = undefined;
-pxt.data.mountVirtualApi("target-config", {
+mountVirtualApi("target-config", {
     getAsync: query => {
         if (!targetConfigPromise)
             targetConfigPromise = pxt.targetConfigAsync()
                 .then(js => {
                     if (js) {
                         pxt.storage.setLocal("targetconfig", JSON.stringify(js))
-                        pxt.data.invalidate("target-config");
-                        pxt.data.invalidate("gh-search");
-                        pxt.data.invalidate("gh-pkgcfg");
+                        invalidate("target-config");
+                        invalidate("gh-search");
+                        invalidate("gh-pkgcfg");
                     }
                     return js;
                 })
@@ -81,11 +322,11 @@ pxt.data.mountVirtualApi("target-config", {
 
 export function invalidateHeader(prefix: string, hd: pxt.workspace.Header) {
     if (hd)
-        pxt.data.invalidate(prefix + ':' + hd.id);
+        invalidate(prefix + ':' + hd.id);
 }
 
-export class Component<TProps, TState> extends React.Component<TProps, TState> implements pxt.data.DataSubscriber {
-    subscriptions: pxt.data.CacheEntry[] = [];
+export class Component<TProps, TState> extends React.Component<TProps, TState> implements DataSubscriber {
+    subscriptions: CacheEntry[] = [];
     renderCoreOk = false;
 
     constructor(props: TProps) {
@@ -101,10 +342,10 @@ export class Component<TProps, TState> extends React.Component<TProps, TState> i
     /**
      * Like getData, but the data is wrapped in a result object that indicates the status of the fetch operation
      */
-    getDataWithStatus<T = any>(path: string): pxt.data.DataFetchResult<T> {
+    getDataWithStatus<T = any>(path: string): DataFetchResult<T> {
         if (!this.renderCoreOk)
             Util.oops("Override renderCore() not render()")
-        return pxt.data.getCached(this, path)
+        return getCached(this, path)
     }
 
     onDataChanged(): void {
@@ -112,7 +353,7 @@ export class Component<TProps, TState> extends React.Component<TProps, TState> i
     }
 
     componentWillUnmount(): void {
-        pxt.data.unsubscribe(this)
+        unsubscribe(this)
     }
 
     child(selector: string) {
@@ -124,7 +365,7 @@ export class Component<TProps, TState> extends React.Component<TProps, TState> i
     }
 
     render() {
-        pxt.data.unsubscribe(this)
+        unsubscribe(this)
         this.renderCoreOk = true;
         return this.renderCore();
     }
@@ -160,4 +401,4 @@ export class PureComponent<TProps, TState> extends React.PureComponent<TProps, T
     }
 }
 
-pxt.data.loadCache();
+loadCache();
